TST: reorganise for FlowProposal refactor
